refactor: move key encode/decode to store package
