+++ packages/store/src/index.js
@@ -56,6 +56,12 @@ export {
   fit,
 } from './patterns/patternMatchers.js';
 export { compareRank, isRankSorted, sortByRank } from './patterns/rankOrder.js';
+export {
+  makeDecodeKey,
+  makeEncodeKey,
+  zeroPad,
+  BIGINT_TAG_LEN,
+} from './patterns/encodeKey.js';
 
 export { makeScalarWeakSetStore } from './stores/scalarWeakSetStore.js';
 export { makeScalarSetStore } from './stores/scalarSetStore.js';
